refactor: fetch 훅을 api 세그먼트로 이동
